Analysis, small updates
