+++ README.md
@@ -1,2 +1,47 @@
 # DEMO-TykeePy
 Public version of TykeePy.
+
+This is a work in progress repository.
+
+Implemented:
+- [x] Bases of Market models
+- [x] Connection to MetaTrader5 and PostgreSQL
+
+Coming soon:
+- [ ] Basic machine learning models
+  - [ ] Linear regression
+  - [ ] Neural networks
+  - [ ] Decision trees
+- [ ] Reinforcement learning model [<sub><sup>(Ugly legacy code)</sup></sub>](https://github.com/refi-refi/ForexBotRL)
+
+Possible future implementations:
+- [ ] Backtesting
+- [ ] Live trading
+
+
+## Installation
+Either use pip to install the package:
+```bash
+pip install .
+```
+or install in editable mode:
+```bash
+pip install -e .
+```
+
+or install using setup.py:
+```bash
+python setup.py install
+```
+
+
+## Disclaimer
+**The information in this repository is for general information only.**
+
+It should not be taken as constituting professional advice from the repository owner - Artūrs Smiltnieks.
+
+Artūrs Smiltnieks is not a financial adviser. You should consider seeking independent legal, financial,
+taxation or other advice to check how the website information relates to your unique circumstances.
+
+Artūrs Smiltnieks is not liable for any loss caused, whether due to negligence or otherwise arising from
+the use of, or reliance on, the information provided directly or indirectly, by use of this repository.
