+++ tests/samples/change_period.py
@@ -1,0 +1,27 @@
+"""
+change_period.py
+Files purpose is to hold a sample data for testing purposes.
+"""
+from pandas import DataFrame
+
+# EURGBP M5 2022-07-01 00:00:00 2022-07-01 00:30:00
+EURGBP_M5 = DataFrame(
+    [
+        (1656633600, 1656633900, 86079, 86104, 85980, 86063, 51, 9),
+        (1656633900, 1656634200, 86062, 86086, 86060, 86074, 51, 17),
+        (1656634200, 1656634500, 86072, 86098, 86063, 86087, 43, 20),
+        (1656634500, 1656634800, 86085, 86116, 86085, 86106, 39, 57),
+        (1656634800, 1656635100, 86105, 86120, 86076, 86110, 38, 79),
+        (1656635100, 1656635400, 86111, 86117, 86089, 86095, 39, 69),
+    ],
+    columns=[
+        "start_ts_utc",
+        "end_ts_utc",
+        "open",
+        "high",
+        "low",
+        "close",
+        "spread",
+        "volume",
+    ],
+)
