+++ tykee/data/__init__.py
@@ -1,2 +1,2 @@
-from .database.db import Database
-from .metatrader.metatrader import MetaTrader
+from tykee.data.sources.database.db import Database
+from tykee.data.sources.metatrader.metatrader import MetaTrader
